fix: New session management
