+++ app/add-ip/page.tsx
@@ -1,12 +1,7 @@
 'use client'
 
 import AddIP from '@/components/AddIP'
-import Authenticated from '@/components/Authenticated'
 
 export default function AddIPPage() {
-  return (
-    <Authenticated requireLit>
-      <AddIP />
-    </Authenticated>
-  )
+  return <AddIP />
 }
