+++ app/details/[id]/page.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import Authenticated from '@/components/Authenticated'
 import DetailIP from '@/components/DetailIP'
 import { useParams } from 'next/navigation'
 
@@ -9,9 +8,5 @@ export default function Token() {
   const docId = params?.id as string // Retrieve the tokenId from the dynamic route
 
   // Require lit for now (in the future this can be on another page)
-  return (
-    <Authenticated requireLit>
-      <DetailIP docId={docId} />
-    </Authenticated>
-  )
+  return <DetailIP docId={docId} />
 }
