+++ app/discovery/[id]/page.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import Authenticated from '@/components/Authenticated'
 import QuestionIP from '@/components/QuestionIP'
 import { useParams } from 'next/navigation'
 
@@ -8,9 +7,5 @@ export default function Token() {
   const params = useParams()
   const docId = params?.id as string // Retrieve the tokenId from the dynamic route
 
-  return (
-    <Authenticated>
-      <QuestionIP docId={docId} />
-    </Authenticated>
-  )
+  return <QuestionIP docId={docId} />
 }
