+++ app/list-ip/page.tsx
@@ -1,12 +1,7 @@
 'use client'
 
-import Authenticated from '@/components/Authenticated'
 import CardGrid from '@/components/card-grid'
 
 export default function Home() {
-  return (
-    <Authenticated requireFirebase={true}>
-      <CardGrid />
-    </Authenticated>
-  )
+  return <CardGrid />
 }
