+++ app/view/[id]/page.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import Authenticated from '@/components/Authenticated'
 import DetailIP from '@/components/DetailIP'
 import { useParams } from 'next/navigation'
 
@@ -8,9 +7,5 @@ export default function Token() {
   const params = useParams()
   const docId = params?.id as string // Retrieve the tokenId from the dynamic route
 
-  return (
-    <Authenticated requireLit>
-      <DetailIP docId={docId} view />
-    </Authenticated>
-  )
+  return <DetailIP docId={docId} view />
 }
