--- components/Authenticated.tsx
@@ -1,465 +1,0 @@
-'use client'
-
-import {
-  createContext,
-  type MouseEvent,
-  type PropsWithChildren,
-  Suspense,
-  useCallback,
-  useEffect,
-  useMemo,
-  useState,
-} from 'react'
-import { AuthModal } from './AuthModal'
-import { useStytch, useStytchUser } from '@stytch/nextjs'
-import Loading from './Loading'
-import {
-  LitAccessControlConditionResource,
-  type LitNodeClient,
-  type AuthMethod,
-  type SessionSigsMap,
-  type AuthSig,
-} from 'lit-wrapper'
-import { publicKeyToAddress } from 'viem/utils'
-import {
-  getAuth,
-  signInWithCustomToken,
-  signOut,
-  type UserCredential,
-} from 'firebase/auth'
-import { usePathname } from 'next/navigation'
-import { useConfig } from '@/app/authLayout'
-import type { Address } from 'viem'
-
-export type Session = {
-  litClient?: LitNodeClient
-  litPromise?: Promise<void>
-  sessionSigs?: {
-    authMethod: AuthMethod
-    pkpPublicKey: string
-    address: Address
-    sessionSigs: SessionSigsMap
-  }
-  delegatedSessionSigs?: (docId: string) => Promise<{
-    sessionSigs: SessionSigsMap
-    capacityDelegationAuthSig: AuthSig
-    address: Address
-  }>
-  fbUser?: UserCredential
-  fbPromise?: Promise<void>
-  isLoggedIn: boolean
-  isLoggingOff: boolean
-  setLoggingOff: (loggingOff: boolean) => void
-}
-
-// Create a singleton instance to persist state across route changes
-let globalSession: Session = {
-  setLoggingOff: () => {},
-  isLoggedIn: false,
-  isLoggingOff: false,
-}
-
-export const sessionContext = createContext<Session>(globalSession)
-
-export default function Authenticated({
-  children,
-  requireLit = false,
-  requireFirebase = true,
-}: PropsWithChildren<{ requireLit?: boolean; requireFirebase?: boolean }>) {
-  const { user, isInitialized } = useStytchUser()
-  const stytchClient = useStytch()
-  const [isLoggingOff, setLoggingOff] = useState(globalSession.isLoggingOff)
-  const [showAuthModal, setShowAuthModal] = useState(false)
-  const config = useConfig()
-  const [, setTrigger] = useState(0)
-
-  globalSession.setLoggingOff = setLoggingOff
-  // Show auth modal if not authenticated and not ignored
-  const amendLoggedIn = useCallback(
-    (state: Session) => {
-      let isLoggedIn = true
-      if (!user) {
-        isLoggedIn = false
-      } else if (requireLit && !state.sessionSigs) {
-        isLoggedIn = false
-      } else if (requireFirebase && !state.fbUser) {
-        isLoggedIn = false
-      }
-      const updatedState = { ...state, isLoggedIn }
-      // Update global session reference
-      globalSession = updatedState
-      return updatedState
-    },
-    [requireLit, requireFirebase, user]
-  )
-
-  useEffect(() => {
-    if (isInitialized && !user) {
-      setShowAuthModal(true)
-    }
-    if (!isInitialized || !user || isLoggingOff) {
-      setLoggingOff(true)
-      try {
-        if (globalSession.litClient) {
-          try {
-            globalSession.litClient.disconnect()
-          } catch (error) {
-            console.error('Error disconnecting Lit client:', error)
-          }
-          globalSession.litClient = undefined
-          globalSession.litPromise = undefined
-          globalSession.sessionSigs = undefined
-          globalSession.delegatedSessionSigs = undefined
-        }
-        if (globalSession.fbUser) {
-          globalSession.fbUser = undefined
-          globalSession.fbPromise = undefined
-          try {
-            signOut(getAuth())
-          } catch (error) {
-            console.error('Error signing out of Firebase:', error)
-          }
-        }
-        globalSession = amendLoggedIn({
-          ...globalSession,
-          isLoggingOff: false,
-          isLoggedIn: false,
-        })
-      } finally {
-        setLoggingOff(false)
-        setTrigger((prev) => prev + 1)
-      }
-    } else {
-      globalSession = amendLoggedIn({
-        ...globalSession,
-      })
-      const task = async () => {
-        const litModule = await import('lit-wrapper')
-        try {
-          if (requireFirebase && !globalSession.fbPromise) {
-            const { currentUser } = getAuth()
-            const fbUser =
-              user &&
-              isInitialized &&
-              currentUser &&
-              currentUser?.uid === user?.user_id
-                ? (currentUser as unknown as UserCredential)
-                : undefined
-            if (fbUser) {
-              globalSession = amendLoggedIn({
-                ...globalSession,
-                fbUser,
-              })
-              setTrigger((prev) => prev + 1)
-              globalSession.fbPromise = Promise.resolve()
-            } else {
-              globalSession.fbPromise = (async () => {
-                true
-                await fetch('/api/exchange', {
-                  headers: {
-                    Authorization: `Bearer ${stytchClient.session.getTokens?.()?.session_jwt}`,
-                  },
-                })
-                  .then((res) => {
-                    if (!res.ok) {
-                      throw new Error('Failed to fetch token')
-                    }
-                    return res.json()
-                  })
-                  .catch((error) => {
-                    console.error('Error fetching token', error)
-                    globalSession.fbPromise = undefined
-                  })
-                  .then((res) => {
-                    return signInWithCustomToken(getAuth(), res.token).then(
-                      (fbUser) => {
-                        globalSession = amendLoggedIn({
-                          ...globalSession,
-                          fbUser,
-                        })
-                        globalSession.fbPromise = Promise.resolve()
-                        setTrigger((prev) => prev + 1)
-                      }
-                    )
-                  })
-                  .catch((error) => {
-                    console.error(error)
-                    globalSession.fbPromise = undefined
-                  })
-              })()
-              globalSession = amendLoggedIn({
-                ...globalSession,
-                fbPromise: globalSession.fbPromise,
-              })
-              setTrigger((prev) => prev + 1)
-            }
-          }
-          if (requireLit && !globalSession.litPromise) {
-            globalSession.litPromise = (async () => {
-              try {
-                const litClient = await litModule.createLitClient({
-                  litNetwork: litModule.LIT_NETWORK.Datil,
-                })
-                globalSession = amendLoggedIn({
-                  ...globalSession,
-                  litClient,
-                })
-                setTrigger((prev) => prev + 1)
-
-                litClient.connect()
-                const { authMethod, provider } = await litModule.authenticate(
-                  litClient,
-                  {
-                    userId: user.user_id,
-                    appId: config.STYTCH_APP_ID as string,
-                    accessToken:
-                      stytchClient.session.getTokens()?.session_jwt || '',
-                    relayApiKey: config.LIT_RELAY_API_KEY as string,
-                  }
-                )
-                // -- setting scope for the auth method
-                // <https://developer.litprotocol.com/v3/sdk/wallets/auth-methods/#auth-method-scopes>
-                const options = {
-                  permittedAuthMethodScopes: [
-                    [litModule.AUTH_METHOD_SCOPE.SignAnything],
-                  ],
-                }
-                let pkps = await provider.fetchPKPsThroughRelayer(authMethod)
-                if (pkps.length <= 0) {
-                  await provider.mintPKPThroughRelayer(authMethod, options)
-                  pkps = await provider.fetchPKPsThroughRelayer(authMethod)
-                }
-                const pkp = pkps[0]
-                const sessionSigs = await litClient
-                  .getPkpSessionSigs({
-                    pkpPublicKey: pkp.publicKey,
-                    chain: 'filecoinCalibrationTestnet',
-                    // capabilityAuthSigs: [capacityDelegationAuthSig],
-                    authMethods: [authMethod],
-                    resourceAbilityRequests: [
-                      {
-                        resource: new litModule.LitPKPResource('*'),
-                        ability: litModule.LIT_ABILITY.PKPSigning,
-                      },
-                      {
-                        resource: new LitAccessControlConditionResource('*'),
-                        ability:
-                          litModule.LIT_ABILITY
-                            .AccessControlConditionDecryption,
-                      },
-                    ],
-                    expiration: new Date(
-                      Date.now() + 1000 * 60 * 10
-                    ).toISOString(), // 10 minutes
-                  })
-                  .catch((error: unknown) => {
-                    console.error(error)
-                    // litClient.removePKPSessionSigs({
-                    //   pkpPublicKey: pkp.publicKey,
-                    //   authMethods: [authMethod],
-                    // })
-                    return litClient.getPkpSessionSigs({
-                      pkpPublicKey: pkp.publicKey,
-                      // litNetwork: litModule.LIT_NETWORK.Datil,
-                      chain: 'filecoinCalibrationTestnet',
-                      // capabilityAuthSigs: [capacityDelegationAuthSig],
-                      authMethods: [authMethod],
-                      resourceAbilityRequests: [
-                        {
-                          resource: new litModule.LitPKPResource('*'),
-                          ability: litModule.LIT_ABILITY.PKPSigning,
-                        },
-                        {
-                          resource: new LitAccessControlConditionResource('*'),
-                          ability:
-                            litModule.LIT_ABILITY
-                              .AccessControlConditionDecryption,
-                        },
-                      ],
-                      expiration: new Date(
-                        Date.now() + 1000 * 60 * 10
-                      ).toISOString(), // 10 minutes
-                    })
-                  })
-                const delegatedSessionSigs = async (docId: string) => {
-                  if (!docId) {
-                    throw new Error('docId is required')
-                  }
-                  const { capacityDelegationAuthSig } = await fetch(
-                    '/api/delegate',
-                    {
-                      method: 'POST',
-                      headers: {
-                        'Content-Type': 'application/json',
-                        Authorization: `Bearer ${stytchClient.session.getTokens()?.session_jwt}`,
-                      },
-                      body: JSON.stringify({
-                        id: docId,
-                        pkp: pkps[0].ethAddress,
-                      }),
-                    }
-                  ).then((res) => {
-                    if (!res.ok) {
-                      throw new Error('Failed to fetch token')
-                    }
-                    return res.json()
-                  })
-
-                  const sessionSigs = await litClient
-                    .getPkpSessionSigs({
-                      pkpPublicKey: pkp.publicKey,
-                      // litNetwork: litModule.LIT_NETWORK.Datil,
-                      chain: 'filecoinCalibrationTestnet',
-                      capabilityAuthSigs: [capacityDelegationAuthSig],
-                      authMethods: [authMethod],
-                      resourceAbilityRequests: [
-                        {
-                          resource: new litModule.LitPKPResource('*'),
-                          ability: litModule.LIT_ABILITY.PKPSigning,
-                        },
-                        {
-                          resource: new LitAccessControlConditionResource('*'),
-                          ability:
-                            litModule.LIT_ABILITY
-                              .AccessControlConditionDecryption,
-                        },
-                      ],
-                      expiration: new Date(
-                        Date.now() + 1000 * 60 * 10
-                      ).toISOString(), // 10 minutes
-                      // capacityDelegationAuthSig,
-                    })
-                    .catch((error: unknown) => {
-                      console.error(error)
-                      // litClient.removePKPSessionSigs({
-                      //   pkpPublicKey: pkp.publicKey,
-                      //   authMethods: [authMethod],
-                      // })
-                      return litClient.getPkpSessionSigs({
-                        pkpPublicKey: pkp.publicKey,
-                        // litNetwork: litModule.LIT_NETWORK.Datil,
-                        chain: 'filecoinCalibrationTestnet',
-                        capabilityAuthSigs: [capacityDelegationAuthSig],
-                        authMethods: [authMethod],
-                        resourceAbilityRequests: [
-                          {
-                            resource: new litModule.LitPKPResource('*'),
-                            ability: litModule.LIT_ABILITY.PKPSigning,
-                          },
-                          {
-                            resource: new LitAccessControlConditionResource(
-                              '*'
-                            ),
-                            ability:
-                              litModule.LIT_ABILITY
-                                .AccessControlConditionDecryption,
-                          },
-                        ],
-                        expiration: new Date(
-                          Date.now() + 1000 * 60 * 10
-                        ).toISOString(), // 10 minutes
-                        // capacityDelegationAuthSig,
-                      })
-                    })
-                  return {
-                    sessionSigs,
-                    capacityDelegationAuthSig,
-                    address: pkp.ethAddress as `0x${string}`,
-                  }
-                }
-                globalSession = amendLoggedIn({
-                  ...globalSession,
-                  litPromise: globalSession.litPromise,
-                  litClient,
-                  delegatedSessionSigs,
-                  sessionSigs: {
-                    authMethod,
-                    pkpPublicKey: pkp.publicKey,
-                    address: publicKeyToAddress(
-                      pkps[0].publicKey as `0x${string}`
-                    ),
-                    sessionSigs,
-                  },
-                })
-                setTrigger((prev) => prev + 1)
-              } catch {
-                globalSession.litPromise = undefined
-              }
-            })()
-            globalSession = amendLoggedIn({
-              ...globalSession,
-              litPromise: globalSession.litPromise,
-            })
-            setTrigger((prev) => prev + 1)
-          }
-        } catch (initError) {
-          console.error('Error initializing Lit client:', initError)
-        }
-      }
-      task()
-    }
-  }, [
-    isInitialized,
-    user,
-    requireLit,
-    requireFirebase,
-    amendLoggedIn,
-    isLoggingOff,
-    config,
-    stytchClient,
-  ])
-
-  // Handle successful authentication
-  const handleAuthSuccess = useCallback(() => {
-    setShowAuthModal(false)
-  }, [])
-
-  const pathname = usePathname()
-  const onClose = useMemo(
-    (event?: MouseEvent<HTMLElement>) => {
-      if (!event) {
-        return () => {
-          setShowAuthModal(false)
-        }
-      }
-      if (pathname === '/') {
-        return () => {
-          setShowAuthModal(false)
-        }
-      }
-      return () => {
-        setShowAuthModal(false)
-        window.location.href = '/'
-      }
-    },
-    [pathname]
-  )
-
-  // When logout is triggered, update global session state
-  useEffect(() => {
-    if (isLoggingOff !== globalSession.isLoggingOff) {
-      globalSession.isLoggingOff = isLoggingOff
-    }
-  }, [isLoggingOff])
-
-  return (
-    <>
-      <sessionContext.Provider value={globalSession}>
-        <Suspense fallback={<Loading />}>
-          {globalSession.isLoggedIn && !globalSession.isLoggingOff ? (
-            children
-          ) : (
-            <Loading />
-          )}
-        </Suspense>
-      </sessionContext.Provider>
-
-      {/* Authentication Modal */}
-      <AuthModal
-        onClose={onClose}
-        isOpen={showAuthModal}
-        onSuccess={handleAuthSuccess}
-      />
-    </>
-  )
-}
