+++ components/DetailIP/ShowDocument.tsx
@@ -1,0 +1,30 @@
+import { Card, CardContent, CardHeader } from '../ui/card'
+import ReactMarkdown from 'react-markdown'
+import remarkGfm from 'remark-gfm'
+import Loading from '../Loading'
+
+type ShowDocumentProps = {
+  viewed?: string
+  view: boolean
+}
+
+export const ShowDocument = ({ viewed, view }: ShowDocumentProps) => {
+  return view ? (
+    <Card className="w-full backdrop-blur-lg bg-background/30 border border-white/10 shadow-xl overflow-hidden">
+      <CardHeader className="pb-4 border-b border-white/10">
+        <div className="flex flex-wrap gap-2 justify-center">
+          <span className="px-3 py-1 text-xs font-medium bg-white/10 text-white/60 rounded-full">
+            Intellectual Property
+          </span>
+        </div>
+      </CardHeader>
+      <CardContent className="pt-6 space-y-5">
+        {viewed ? (
+          <ReactMarkdown remarkPlugins={[remarkGfm]}>{viewed}</ReactMarkdown>
+        ) : (
+          <Loading text="Decrypting Content" />
+        )}
+      </CardContent>
+    </Card>
+  ) : null
+}
