+++ hooks/useProducts.tsx
@@ -6,6 +6,8 @@ import {
   where,
 } from 'firebase/firestore'
 import { useEffect, useState } from 'react'
+import { useDebounce } from 'use-debounce'
+import { handleError } from './useIP'
 
 export type Price = {
   id: string
@@ -58,51 +60,67 @@ export function useProducts(): Record<string, Product> {
     const prices = collectionGroup(fb, 'prices')
     const activePrices = query(prices, where('active', '==', true))
     const shutdown: Map<[string, string], () => void> = new Map()
-    const close = onSnapshot(activePrices, async (priceSnap) => {
-      for (const priceDoc of priceSnap.docs) {
-        // We're already listening on the product doc
-        const productId = priceDoc.ref.parent?.parent?.id || ''
-        if (shutdown.has(['product', productId])) {
-          return
-        }
-        setProducts((prev) => ({
-          ...prev,
-          [productId]: {
-            ...prev[priceDoc.ref.parent.id],
-            id: productId,
-            prices: {
-              ...(prev[productId]?.prices || {}),
-              [priceDoc.id]: {
-                id: priceDoc.id,
-                ...priceDoc.data(),
-              } as Price,
-            } as Record<string, Price>,
-          } as Product,
-        }))
-        const parentRef = priceDoc.ref.parent?.parent
-        if (!parentRef) {
-          return
-        }
-        const close = onSnapshot(parentRef, (productSnap) => {
-          const data = productSnap.data()
-          if (!data?.active) {
+    const close = onSnapshot(
+      activePrices,
+      async (priceSnap) => {
+        for (const priceDoc of priceSnap.docs) {
+          // We're already listening on the product doc
+          const productId = priceDoc.ref.parent?.parent?.id || ''
+          if (shutdown.has(['product', productId])) {
             return
           }
-          setProducts((prev) => ({
-            ...prev,
-            [productSnap.id]: {
-              id: productSnap.id,
-              prices: (prev[productSnap.id]?.prices || {}) as Record<
-                string,
-                Price
-              >,
-              ...data,
-            } as Product,
-          }))
-        })
-        shutdown.set(['product', productId], close)
-      }
-    })
+          setProducts((prev) => {
+            const data = {
+              ...prev,
+              [productId]: {
+                ...prev[priceDoc.ref.parent.id],
+                id: productId,
+                prices: {
+                  ...(prev[productId]?.prices || {}),
+                  [priceDoc.id]: {
+                    id: priceDoc.id,
+                    ...priceDoc.data(),
+                  } as Price,
+                } as Record<string, Price>,
+              } as Product,
+            }
+            // console.log('Product data', data)
+            return data
+          })
+          const parentRef = priceDoc.ref.parent?.parent
+          if (!parentRef) {
+            return
+          }
+          const close = onSnapshot(
+            parentRef,
+            (productSnap) => {
+              const data = productSnap.data()
+              if (!data?.active) {
+                return
+              }
+              setProducts((prev) => {
+                const output = {
+                  ...prev,
+                  [productSnap.id]: {
+                    id: productSnap.id,
+                    prices: (prev[productSnap.id]?.prices || {}) as Record<
+                      string,
+                      Price
+                    >,
+                    ...data,
+                  } as Product,
+                }
+                // console.log('Product data', output)
+                return output
+              })
+            },
+            handleError(parentRef)
+          )
+          shutdown.set(['product', productId], close)
+        }
+      },
+      handleError('collectionGroup prices')
+    )
     shutdown.set(['prices', ''], close)
     return () => {
       for (const close of shutdown.values()) {
@@ -110,5 +128,5 @@ export function useProducts(): Record<string, Product> {
       }
     }
   }, [])
-  return products
+  return useDebounce(products, 500)[0] as Record<string, Product>
 }
